CR-8301 AME-8480 Add release profile for web agent docs
This patch completes addition of a release profile for web agent docs
suitable for use with the Maven release plugin.

Most of the changes were committed with r15961.

The output at this time is only the release-ready BackStage input.

Once we put together a Jenkins job for the doc release,
we can determine whether it makes more sense to drive the commit to the BackStage git repo
from Maven or from a script in the Jenkins job.


git-svn-id: https://svn.forgerock.org/openam/trunk@15964 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d